BFS implementation is ready.
This commit includes the implementation itself.

Also, I've had a pleasant experience modifing the internal data
structure of the BFSNode (from HashMap to Set) because I followed
the law of demeter, so, there was no side effects on my code
(well, it was, but because I changed a collection for an array
as a return, but that's more a design change).